+++ da/weavevm/signer/private_key_signer.go
@@ -9,7 +9,7 @@ import (
 	"math/big"
 	"strings"
 
-	weaveVMtypes "github.com/dymensionxyz/dymint/da/weave_vm/types"
+	weaveVMtypes "github.com/dymensionxyz/dymint/da/weavevm/types"
 	"github.com/ethereum/go-ethereum/common"
 	"github.com/ethereum/go-ethereum/common/hexutil"
 	ethtypes "github.com/ethereum/go-ethereum/core/types"
